Update handling of userMoney check
Update the way that userMoney is checked so that the user is not prompted to "Press 1 to purchase another: " if they don't have enough money to buy another drink. If they don't, will simply let the user know and thank them for their purchase. Also added '$' to any out put of the user's money